+++ src/test/java/hu/kazocsaba/imageviewer/gui/PopupTest.java
@@ -8,12 +8,9 @@ import javax.swing.SwingUtilities;
 import hu.kazocsaba.imageviewer.ImageViewer;
 import org.fest.swing.edt.GuiActionRunner;
 import org.fest.swing.edt.GuiQuery;
-import org.fest.swing.edt.GuiTask;
 import org.fest.swing.exception.ComponentLookupException;
 import org.fest.swing.fixture.FrameFixture;
-import org.fest.swing.fixture.JPopupMenuFixture;
 import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
@@ -26,9 +23,17 @@ public class PopupTest {
 	private final Dimension imageSize=new Dimension(100, 100);
 	private FrameFixture window;
 	private ImageViewer viewer;
-
-	@Before
-	public void setUp() {
+	
+	@After
+	public void tearDown() {
+		if (window!=null) {
+			window.cleanUp();
+			window=null;
+		}
+	}
+	
+	@Test
+	public void testDefault() {
 		JFrame frame = GuiActionRunner.execute(new GuiQuery<JFrame>() {
 
 			@Override
@@ -41,25 +46,26 @@ public class PopupTest {
 		});
 		window = new FrameFixture(frame);
 		window.show(windowSize);
+		
+		JViewport viewport=GuiUtils.getScrollPaneFor(viewer).getViewport();
+		window.showPopupMenuAt(SwingUtilities.convertPoint(viewport, viewport.getLocation(), window.component()));
 	}
-
-	@After
-	public void tearDown() {
-		window.cleanUp();
-	}
-	
 	@Test
 	public void testDisable() {
-		JViewport viewport=GuiUtils.getScrollPaneFor(viewer).getViewport();
-		final JPopupMenuFixture popup = window.showPopupMenuAt(SwingUtilities.convertPoint(viewport, viewport.getLocation(), window.component()));
-		GuiActionRunner.execute(new GuiTask() {
+		JFrame frame = GuiActionRunner.execute(new GuiQuery<JFrame>() {
 
 			@Override
-			protected void executeInEDT() throws Throwable {
-				popup.component().setVisible(false);
-				viewer.setPopupMenu(null);
+			protected JFrame executeInEDT() throws Throwable {
+				JFrame frame = new JFrame("Test window");
+				viewer = new ImageViewer(new BufferedImage(imageSize.width, imageSize.height, BufferedImage.TYPE_BYTE_GRAY), false);
+				frame.add(viewer.getComponent());
+				return frame;
 			}
 		});
+		window = new FrameFixture(frame);
+		window.show(windowSize);
+		
+		JViewport viewport=GuiUtils.getScrollPaneFor(viewer).getViewport();
 		try {
 			window.showPopupMenuAt(SwingUtilities.convertPoint(viewport, viewport.getLocation(), window.component()));
 			fail("Popup shown");